chapters auth done!

+++ src/use-cases/chapters/delete-chapter.ts
@@ -2,10 +2,12 @@ import { Chapter } from "@prisma/client";
 import { ChaptersRepository } from "@/repositories/chapters-repository";
 import { SubscriptionsRepository } from "@/repositories/subscriptions-repository";
 import { ResourceNotFoundError } from "../errors/resource-not-found-error";
+import { ForbiddenError } from "../errors/forbidden-error";
 
 interface DeleteChapterUseCaseRequest {
     subscriptionId: string;
     number: number;
+    userId: string;
 }
 
 interface DeleteChapterUseCaseResponse {
@@ -18,11 +20,15 @@ export class DeleteChapterUseCase {
         private subscriptionsRepository: SubscriptionsRepository
     ) {}
 
-    async execute({ subscriptionId, number }: DeleteChapterUseCaseRequest): Promise<DeleteChapterUseCaseResponse> {
+    async execute({ subscriptionId, number, userId }: DeleteChapterUseCaseRequest): Promise<DeleteChapterUseCaseResponse> {
         const subscription = await this.subscriptionsRepository.getSubscriptionById({ subscriptionId });
 
         if (!subscription) {
             throw new ResourceNotFoundError();
+        }
+
+        if (subscription.userId !== userId) {
+            throw new ForbiddenError();
         }
 
         const chapter = await this.chaptersRepository.delete({
